Issue-9767 - Extract array_dims, array_ndims and flatten functions from functions-array subcrate' s kernels and udf containers
