+++ datafusion/functions-array/src/lib.rs
@@ -32,10 +32,11 @@ mod array_has;
 mod cardinality;
 mod concat;
 mod core;
+mod dimension;
 mod empty;
 mod except;
 mod extract;
-mod kernels;
+mod flatten;
 mod length;
 mod position;
 mod range;
@@ -48,7 +49,6 @@ mod rewrite;
 mod set_ops;
 mod sort;
 mod string;
-mod udf;
 mod utils;
 
 use datafusion_common::Result;
@@ -67,12 +67,15 @@ pub mod expr_fn {
     pub use super::concat::array_concat;
     pub use super::concat::array_prepend;
     pub use super::core::make_array;
+    pub use super::dimension::array_dims;
+    pub use super::dimension::array_ndims;
     pub use super::empty::array_empty;
     pub use super::except::array_except;
     pub use super::extract::array_element;
     pub use super::extract::array_pop_back;
     pub use super::extract::array_pop_front;
     pub use super::extract::array_slice;
+    pub use super::flatten::flatten;
     pub use super::length::array_length;
     pub use super::position::array_position;
     pub use super::position::array_positions;
@@ -93,9 +96,6 @@ pub mod expr_fn {
     pub use super::sort::array_sort;
     pub use super::string::array_to_string;
     pub use super::string::string_to_array;
-    pub use super::udf::array_dims;
-    pub use super::udf::array_ndims;
-    pub use super::udf::flatten;
 }
 
 /// Registers all enabled packages with a [`FunctionRegistry`]
@@ -105,9 +105,9 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         string::string_to_array_udf(),
         range::range_udf(),
         range::gen_series_udf(),
-        udf::array_dims_udf(),
+        dimension::array_dims_udf(),
         cardinality::cardinality_udf(),
-        udf::array_ndims_udf(),
+        dimension::array_ndims_udf(),
         concat::array_append_udf(),
         concat::array_prepend_udf(),
         concat::array_concat_udf(),
@@ -122,7 +122,7 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         array_has::array_has_any_udf(),
         empty::array_empty_udf(),
         length::array_length_udf(),
-        udf::flatten_udf(),
+        flatten::flatten_udf(),
         sort::array_sort_udf(),
         repeat::array_repeat_udf(),
         resize::array_resize_udf(),
